% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_style.R
\name{my_style}
\alias{my_style}
\title{Add mytheme to ggplot chart}
\usage{
my_style()
}
\description{
This function allows you to add the andy theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
my_style()
}
\keyword{my_style}
