% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme1.R
\name{Theme}
\alias{Theme}
\title{Create a theme experiment instance for a ggplot}
\usage{
Theme(ggObject)
}
\arguments{
\item{ggObject}{A ggplot class object.}
}
\value{
An instance of environment
}
\description{
Create a theme experiment instance for a ggplot
}
\examples{
none
}
