% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_paper.R
\name{theme_paper}
\alias{theme_paper}
\title{Theme of ggplot2 for scientific paper usage}
\usage{
theme_paper()
}
\description{
Theme of ggplot2 for scientific paper usage.
No background panel.
No grid of plot area.
The color of all texts are "black".
}
\examples{
theme_paper()
}
\author{
Ken Takasawa
}
