% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andy_style.R
\name{andy_style}
\alias{andy_style}
\title{Add andy theme to ggplot chart}
\usage{
andy_style()
}
\description{
This function allows you to add the andy theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
andy_style()
}
\keyword{andy_style}
